requirements.txt
pyrogram
requests
flask
python-dotenv
tgcrypto
pyrofork
